BookmarkController migrating to JDBC!

+++ server/src/main/java/dev/findfirst/core/repository/jdbc/BookmarkTagRepositoryCustomImpl.java
@@ -1,25 +1,29 @@
 package dev.findfirst.core.repository.jdbc;
 
 import dev.findfirst.core.model.jdbc.BookmarkTag;
+import dev.findfirst.security.userAuth.tenant.contexts.TenantContext;
+import lombok.RequiredArgsConstructor;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 
 @Repository
+@RequiredArgsConstructor
 public class BookmarkTagRepositoryCustomImpl implements BookmarkTagRepositoryCustom {
-  @Autowired
-  private JdbcTemplate jdbcTemplate;
+  private final JdbcTemplate jdbcTemplate;
+
+  private final TenantContext tContext;
+
 
   @Override
-  public int update(BookmarkTag bookmarkTag) {
+  public int saveBookmarkTag(BookmarkTag bookmarkTag) {
     String sql = "INSERT INTO bookmark_tag VALUES(?, ?)";
     return jdbcTemplate.update(sql, bookmarkTag.getBookmarkId(), bookmarkTag.getTagId());
   }
 
   @Override
   public int deleteBookmarkTag(BookmarkTag bookmarkTag) {
-    String sql = "DELETE FROM bookmark_tag bt WHERE bt.bookmark_id = ? AND bt.tag_id = ?";
-    return jdbcTemplate.update(sql, bookmarkTag.getBookmarkId(), bookmarkTag.getTagId());
+    String sql = "DELETE FROM bookmark_tag bt USING bookmark WHERE bt.bookmark_id = ? AND bt.tag_id = ? AND bookmark.tenant_id = ?";
+    return jdbcTemplate.update(sql, bookmarkTag.getBookmarkId(), bookmarkTag.getTagId(), tContext.getTenantId());
   }
 }
